Working Shop UI
The shop UI appears when the user presses P. This is planned to only work when the user approaches a shop but thats yet to be programmed. The shop subtracts money and disables gear from being bought again but the items don't have functionality yet.

Fixes issues with scene loading and minor balancing tweaks.